backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic[email]
boto3
python-dotenv
